Network_Recon/requirements.txt
whois
dnspython
requests
beautifulsoup4